prepare for Python3

+++ debugmotion.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 # vim: set et sw=4 sts=4 fileencoding=utf-8:
 from time import sleep,time
 #from argparse import ArgumentParser
@@ -75,7 +75,7 @@ def main(fobj=None,width=1280,height=960,video=False):
 
     caption = 'piCAMTracker::cv2'
     camera = faked_camera(resx=width, resy=height)
-    image = np.ones((height/2,width/2,3), np.uint8) * 220
+    image = np.ones((int(height/2),int(width/2),3), np.uint8) * 220
     rot_img = np.flipud(np.rot90(np.flipud(image),k=1))
     tracker = picamtracker.MotionTracker.Tracker(camera, greenLed=None, redLed=None, config=config)
     tracker.setup_sizes(height/16, width/16)
